FIX: 파일구조 오류 수정 과정 ( 패키지 내임 수정)
