+++ src/main/java/swengineering/team7/issuemanagementsystem/DTOs/IssueCountByDateDTO.java
@@ -1,4 +1,4 @@
-package swengineering.team7.issuemanagementsystem.DTO;
+package swengineering.team7.issuemanagementsystem.DTOs;
 
 import java.time.LocalDate;
