--- src/main/java/swengineering/team7/issuemanagementsystem/dto/ProjectAssignedUserDTO.java
@@ -1,23 +1,0 @@
-package swengineering.team7.issuemanagementsystem.DTO;
-
-import java.util.List;
-
-public class ProjectAssignedUserDTO {
-
-    private Long projectId;
-    private List<String> assignedUsersID;
-
-    public ProjectAssignedUserDTO(Long projectId, List<String> assignedUsersID) {
-        this.projectId = projectId;
-        this.assignedUsersID = assignedUsersID;
-    }
-
-    public Long getProjectId() {
-        return projectId;
-    }
-
-    public List<String> getAssignedUsersID() {
-        return assignedUsersID;
-    }
-
-}
