+++ src/main/java/swengineering/team7/issuemanagementsystem/entity/Tester.java
@@ -1,6 +1,7 @@
 package swengineering.team7.issuemanagementsystem.entity;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "Tester")
