+++ src/main/java/swengineering/team7/issuemanagementsystem/service/CommentService.java
@@ -1,16 +1,11 @@
 package swengineering.team7.issuemanagementsystem.service;
 
-import swengineering.team7.issuemanagementsystem.DTO.IssueDTO;
+import swengineering.team7.issuemanagementsystem.DTOs.IssueDTO;
 import swengineering.team7.issuemanagementsystem.dto.ProjectDTO;
-import swengineering.team7.issuemanagementsystem.DTO.UserInformationDTO;
 import swengineering.team7.issuemanagementsystem.entity.Comment;
 import swengineering.team7.issuemanagementsystem.entity.Issue;
-import swengineering.team7.issuemanagementsystem.entity.Tester;
-import swengineering.team7.issuemanagementsystem.entity.Project;
 import swengineering.team7.issuemanagementsystem.entity.User;
-import swengineering.team7.issuemanagementsystem.exception.NoBelong;
-import swengineering.team7.issuemanagementsystem.DTO.CommentDTO;
-import swengineering.team7.issuemanagementsystem.entity.Issue;
+import swengineering.team7.issuemanagementsystem.DTOs.CommentDTO;
 
 import swengineering.team7.issuemanagementsystem.repository.CommentRepository;
 import swengineering.team7.issuemanagementsystem.repository.IssueRepository;
