+++ src/main/java/swengineering/team7/issuemanagementsystem/service/IssueService.java
@@ -1,13 +1,11 @@
 package swengineering.team7.issuemanagementsystem.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
-import swengineering.team7.issuemanagementsystem.DTO.*;
 import swengineering.team7.issuemanagementsystem.entity.*;
 import swengineering.team7.issuemanagementsystem.repository.CommentRepository;
-import swengineering.team7.issuemanagementsystem.DTO.IssueDTO;
+import swengineering.team7.issuemanagementsystem.DTOs.IssueDTO;
 import swengineering.team7.issuemanagementsystem.dto.SearchInfoDTO;
 import swengineering.team7.issuemanagementsystem.entity.Issue;
 import swengineering.team7.issuemanagementsystem.entity.Project;
@@ -19,7 +17,6 @@ import swengineering.team7.issuemanagementsystem.repository.UserRepository;
 import swengineering.team7.issuemanagementsystem.util.Priority;
 import swengineering.team7.issuemanagementsystem.util.SearchType;
 
-import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
